Remove a bunch of duplicated tests
There were "unit" and "integration" tests, which were mostly completely duplicated. With the only exception that the "integration" tests were calling the constructor, and the "unit" test just sets private fields directly.
